+++ v8/app.js
@@ -19,7 +19,9 @@ mongoose.connect("mongodb://user:testapp@ds129050.mlab.com:29050/ecdata");
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(express.static(__dirname + "/public"));
 app.set("view engine", "ejs");
-seedDB();
+
+//seed the database
+// seedDB();
 
 // ======================
 // PASSPORT CONFIGURATION