+++ v8/models/mComment.js
@@ -3,7 +3,13 @@ const mongoose = require("mongoose");
 // === Schema Setup ===
 const commentSchema = mongoose.Schema({
     text: String,
-    author: String,
+    author: {
+        id: {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: "User"
+        },
+        username: String
+    }
 });
 
 // === Model Setup & Export ===